FlowerDelivery/requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
magic-filter==1.0.12
multidict==6.0.5
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-telegram-bot==21.4
pytz==2024.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
